Rename backend folder to dss
