--- pkg/backend/server.go
@@ -1,45 +1,0 @@
-package backend
-
-import (
-	"context"
-
-	dspb "github.com/steeling/InterUSS-Platform/pkg/dssproto"
-)
-
-type DSSStorageInterface interface{}
-
-type DSSServer struct {
-	store DSSStorageInterface
-}
-
-func NewServer() (*DSSServer, error) {
-	return &DSSServer{}, nil
-}
-
-func (s *DSSServer) DeleteIdentificationServiceArea(ctx context.Context, req *dspb.DeleteIdentificationServiceAreaRequest) (*dspb.DeleteIdentificationServiceAreaResponse, error) {
-	return nil, nil
-}
-
-func (s *DSSServer) DeleteSubscription(ctx context.Context, req *dspb.DeleteSubscriptionRequest) (*dspb.DeleteSubscriptionResponse, error) {
-	return nil, nil
-}
-
-func (s *DSSServer) SearchIdentificationServiceAreas(ctx context.Context, req *dspb.SearchIdentificationServiceAreasRequest) (*dspb.SearchIdentificationServiceAreasResponse, error) {
-	return nil, nil
-}
-
-func (s *DSSServer) SearchSubscriptions(ctx context.Context, req *dspb.SearchSubscriptionsRequest) (*dspb.SearchSubscriptionsResponse, error) {
-	return nil, nil
-}
-
-func (s *DSSServer) GetSubscription(ctx context.Context, req *dspb.GetSubscriptionRequest) (*dspb.GetSubscriptionResponse, error) {
-	return nil, nil
-}
-
-func (s *DSSServer) PutIdentificationServiceArea(ctx context.Context, req *dspb.PutIdentificationServiceAreaRequest) (*dspb.PutIdentificationServiceAreaResponse, error) {
-	return nil, nil
-}
-
-func (s *DSSServer) PutSubscription(ctx context.Context, req *dspb.PutSubscriptionRequest) (*dspb.PutSubscriptionResponse, error) {
-	return nil, nil
-}
